% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{apex_treemap}
\alias{apex_treemap}
\title{Apexchart Treemap}
\usage{
apex_treemap(series = NULL, colors = NULL, legend_size = 15)
}
\arguments{
\item{series}{The series to plot. Generated from `jsonify_metadata`.}

\item{colors}{Treemap fill colors.}

\item{legend_size}{Legend font size.}
}
\value{
An apexchart interactive plot.
}
\description{
Generate a treemap using the `apexchart` library (see \url{https://dreamrs.github.io/apexcharter/index.html})
}
\examples{
\dontrun{
palette <- c(
  "#440154", "#30678D", "#35B778",
  "#FDE725", "#FCA35D", "#D32F2F", "#67001F"
)
colors <- palette \%>\% strtrim(width = 7)
dat <- jsonify_metadata()
apex_treemap(series = dat, colors = colors, legend_size = 15)
}
}
