% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-dataverse-medata.R
\name{clean_dataverse_metadata}
\alias{clean_dataverse_metadata}
\title{Clean metadata raw files}
\usage{
clean_dataverse_metadata(file_path = NULL)
}
\arguments{
\item{file_path}{Path of raw metadata file.}
}
\value{
A dataframe.
}
\description{
Read and clean Dataverse raw metadata files.
}
