% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse_metadata.R
\docType{data}
\name{dataverse_metadata}
\alias{dataverse_metadata}
\title{Dataverse Metadata}
\format{
A dataframe
}
\usage{
dataverse_metadata
}
\description{
A dataset containing metadata information of WorldFish data on the Dataverse
portal.
}
\keyword{datasets}
