% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing-functions.R
\name{filterby_word}
\alias{filterby_word}
\title{Filter metadata by word}
\usage{
filterby_word(word = NULL)
}
\arguments{
\item{word}{Character string to be be filtered.}
}
\value{
A dataframe
}
\description{
Filter metadata by a specific word in dataset title and/or dataset keyword list.
}
