% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataverse-data.R
\name{get_dataset}
\alias{get_dataset}
\title{Get dataset metadata}
\usage{
get_dataset(doi = NULL, dataverse_key = NULL)
}
\arguments{
\item{doi}{The dataset DOI (eg. "10.7910/DVN/EXSAJ7").}

\item{dataverse_key}{Dataverse token.}
}
\value{
A dataframe.
}
\description{
A wrapper function of [dataverse::get_dataset].
}
\details{
This function returns dataset' metadata information.
}
\examples{
\dontrun{
get_dataset(doi = "10.7910/DVN/WMLQ3D", token = DATAVERSE_TOKEN)
}
}
