% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataverse-data.R
\name{get_dataset_file}
\alias{get_dataset_file}
\title{Get file from dataverse dataset}
\usage{
get_dataset_file(dataset = NULL, file_id = NULL)
}
\arguments{
\item{dataset}{A dataverse dataset returned from [aquadata.data.mappind::get_dataset].}

\item{file_id}{The dataverse file id.}
}
\value{
An object based on the file extension. A dataframe in case of .csv or .xlsx,
 a text for .pdf and .docx files.
}
\description{
A wrapper function of [dataverse::get_dataframe_by_id].
}
\details{
This function returns a specific file associated to a dataverse dataset.
It returns an object based on the extension of the file. A dataframe in case of
.csv or .xlsx, a text for .pdf and .docx files.
}
\examples{
\dontrun{
dataverse_dataset <- get_dataset(doi = "10.7910/DVN/WMLQ3D", token = DATAVERSE_TOKEN)
get_dataset_file(dataset = dataverse_dataset, id = 4570239)
}
}
