% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataverse-data.R
\name{get_dataset_insights}
\alias{get_dataset_insights}
\title{Get dataverse dataset insights}
\usage{
get_dataset_insights(parent_table = NULL, index = NULL)
}
\arguments{
\item{parent_table}{A unique dataverse dataset.}

\item{index}{A number indicating the row number of the `parent_table`.}
}
\value{
A dataframe showing the objects included in the selected dataset.
}
\description{
This function is helpful when exploring the content of a dataverse object
from the the metadata table [aquadata.data.mapping::dataverse_metadata].
}
\examples{
\dontrun{
# Select a random dataverse dataset
dataset <- aquadata.data.mapping::dataverse_metadata \%>\% dplyr::slice(5)
get_dataverse_metadata(
  parent_table = aquadata.data.mapping::dataverse_metadata,
  index = 5
)
}
}
