% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataverse-data.R
\name{get_dataverse_metadata}
\alias{get_dataverse_metadata}
\title{Download dataverse metadata}
\usage{
get_dataverse_metadata(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See [logger::log_levels] for more
details}
}
\value{
Nothing, this function is used for its side effects.
}
\description{
This function uses [purrr::walk] together with `get_organization_metadata`
to download Dataverse metadata from several organizations. Data are then
stored as csv files into data-raw folder.
}
\examples{
\dontrun{
get_dataverse_metadata()
}
}
