% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataverse-data.R
\name{get_organization_metadata}
\alias{get_organization_metadata}
\title{Download metadata}
\usage{
get_organization_metadata(organization = NULL)
}
\arguments{
\item{organization}{The Dataverse alias of the entity posting the data.}
}
\value{
A csv with metadata information.
}
\description{
This function is a wrapper of a python script `get_dataverse_metadata.py` udeful
to download metadata from a defined Dataverse.
}
\examples{
\dontrun{
get_organization_metadata(organization = "CIAT")
}
}
