% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataverse-data.R
\name{get_organization_metadata_test}
\alias{get_organization_metadata_test}
\title{Download metadata test func}
\usage{
get_organization_metadata_test()
}
\value{
consectetur adipiscing elit
}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit
}
\examples{
\dontrun{
get_organization_metadata()
}
}
