% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guestbook_responses.R
\docType{data}
\name{guestbook_responses}
\alias{guestbook_responses}
\title{Guest responses}
\format{
A dataframe
}
\usage{
guestbook_responses
}
\description{
A dataset containing user activity responses on WordFish data from the Dataverse
portal.
}
\keyword{datasets}
