% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{jsonify_metadata}
\alias{jsonify_metadata}
\title{Rearrange metadata for apexcharter}
\usage{
jsonify_metadata()
}
\value{
A list.
}
\description{
This function is strictly related to `apexcharter`, as it rearrange the
metadata dataframe in the shape of "list of lists" in order to be correctly
processed by `apex_treemap` function.
}
\examples{
\dontrun{
metadata_lists <- jsonify_metadata()
}
}
