% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-dataverse-medata.R
\name{process_raw_metadata}
\alias{process_raw_metadata}
\title{Clean raw metadata}
\usage{
process_raw_metadata(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache logj4) log level used as a
threshold for the logging infrastructure. See [logger::log_levels] for more
details}
}
\value{
Nothing, this function clean and store raw metadata into "data" folder.
}
\description{
This function uses `clean_metadata_files` to clean metadata raw files into an
unique and structured .rda file.
}
\examples{
\dontrun{
process_raw_metadata()
}
}
