% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt-wrapper.R
\name{qa_bot_wrapper}
\alias{qa_bot_wrapper}
\title{Q&A wrapper}
\usage{
qa_bot_wrapper(
  document_path = NULL,
  openaikey = NULL,
  temperature = 0.5,
  query_text = NULL
)
}
\arguments{
\item{document_path}{The path of the txt document to summarise.}

\item{openaikey}{The OpenAI key token.}

\item{temperature}{Level of randomness or "creativity" in the generated text (see \url{https://platform.openai.com/docs/api-reference/completions/create}).}

\item{query_text}{The query to interact with the document.}
}
\value{
A text answer.
}
\description{
This function is wrapper of LangChain Q&A module for pdfs documents.
}
\examples{
\dontrun{
qa_bot_wrapper(
  document_path = "mypath/dummy_file.pdf",
  openaikey = OPENAI_KEY,
  temperature = 0.7,
  query_text = "What are the most important points of the document?"
)
}
}
