% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shiny_server}
\alias{shiny_server}
\title{The Shiny App Server.}
\usage{
shiny_server(input, output)
}
\arguments{
\item{input}{input set by Shiny.}

\item{output}{output set by Shiny.}
}
\description{
The Shiny App Server.
}
