% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{shiny_ui}
\alias{shiny_ui}
\title{The Shiny App UI.}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
shiny_ui
}
\description{
The Shiny App UI.
}
\keyword{datasets}
