% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt-wraps.R
\name{summarise_chatgpt_wrapper}
\alias{summarise_chatgpt_wrapper}
\title{Summarise document using LangChain and ChatGPT}
\usage{
summarise_chatgpt_wrapper(document_path, openaikey, engine, temperature)
}
\arguments{
\item{document_path}{The path of the txt document to summarise.}

\item{openaikey}{The OpenAI key token.}

\item{engine}{ChatGPT model (see \url{https://platform.openai.com/docs/models}).}

\item{temperature}{Level of randomness or "creativity" in the generated text (see \url{https://platform.openai.com/docs/api-reference/completions/create})}
}
\value{
A summarised text.
}
\description{
This function is a R wrapper of `chatgpt_chaindoc.py`, a python function aimed
to summarise large text documents using LangChain and ChatGPT.
}
