% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfish_guestbook_responses.R
\docType{data}
\name{worldfish_guestbook_responses}
\alias{worldfish_guestbook_responses}
\title{Worldfish Guest responses}
\format{
A dataframe
}
\usage{
worldfish_guestbook_responses
}
\description{
A dataset containing user activity responses on WordFish data from the Dataverse
portal.
}
\keyword{datasets}
