% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-dataverse-medata.R
\name{clean_metadata_files}
\alias{clean_metadata_files}
\title{Clean metadata raw files}
\usage{
clean_metadata_files(folder_path = NULL, file_path = NULL)
}
\arguments{
\item{folder_path}{Path of folder where raw metadata are located.}

\item{file_path}{Path of raw metadata file.}
}
\value{
A dataframe.
}
\description{
Read and clean Dataverse raw metadata files.
}
