% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataverse-data.R
\name{get_dataset_insights}
\alias{get_dataset_insights}
\title{Get dataverse dataset insights}
\usage{
get_dataset_insights(selected_dataset)
}
\arguments{
\item{selected_dataset}{A unique dataverse dataset.}
}
\value{
A dataframe showing the objects included in the selected dataset.
}
\description{
This function is helpful when exploring the content of a dataverse object
from the the metadata table [aquadata.data.mapping::dataverse_metadata].
}
\examples{
\dontrun{
# Select a random dataverse dataset
dataset <- aquadata.data.mapping::dataverse_metadata \%>\% dplyr::slice(5)
get_dataverse_metadata(selected_dataset = dataset)
}
}
