% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataverse-data.R
\name{get_organization_metadata_test}
\alias{get_organization_metadata_test}
\title{Download metadata test}
\usage{
get_organization_metadata_test()
}
\value{
Nothing
}
\description{
Testing function to get python files path
}
\examples{
\dontrun{
get_organization_metadata(organization = "CIAT")
}
}
