% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt-wraps.R
\name{summarise_chatgpt_wrapper}
\alias{summarise_chatgpt_wrapper}
\title{Summarise document using LangChain and ChatGPT}
\usage{
summarise_chatgpt_wrapper(
  document_path,
  openaikey,
  engine,
  temperature,
  refine_text
)
}
\arguments{
\item{document_path}{The path of the txt document to summarise.}

\item{openaikey}{The OpenAI key token.}

\item{engine}{ChatGPT model (see \url{https://platform.openai.com/docs/models}).}

\item{temperature}{Level of randomness or "creativity" in the generated text (see \url{https://platform.openai.com/docs/api-reference/completions/create}).}

\item{refine_text}{A text useful to refine the original prompt (see \url{https://langchain.readthedocs.io/en/latest/modules/indexes/chain_examples/summarize.html#the-refine-chain} for details).}
}
\value{
A summarised text.
}
\description{
This function is a R wrapper of `chatgpt_chaindoc.py`, a python script function
aimed to elaborate large text documents using LangChain and ChatGPT.
`chatgpt_chaindoc.py` uses the refine chain method to elaborate the text
(see \url{https://langchain.readthedocs.io/en/latest/modules/indexes/chain_examples/summarize.html#the-refine-chain} for details).
}
\examples{
\dontrun{
output <-
summarise_chatgpt_wrapper(
document_path = "inst/docs_dataverse/5636634.txt",
openaikey = OPENAI_KEY,
engine = "gpt-3.5-turbo",
temperature = 0.7,
refine_text = pars$openai$refine_prompts$p1
)
cat(output$output_text)
}
}
